split up ecolor crate in lots of modules
